Created widget for main screen
NOTE : This widget may need to be removed, as there is no "main" screen idea as of now. Maybe this could be a "greeting" or "what's new" or "MOTD" screen of sorts. The idea would be to have a screen for each tab in the program, so this is low priority.