Separated parent and child padding into 2 separate variables
Note: Maybe it would make more sense to only have a single padding bar and add it to all of the Mpup widget classes. That padding should be refered to the padding between the widget and its parent (what is known here as "paddingParent"). That way, if the child object needs padding, it provides it itself.