Modified settings menu to load settings every time the menu is loaded
This change allows users to edit their settings while the program is opened by using an external text editor or some other program that managers settings. In the event that the settings were to be externally modified, they would be properly reloaded when entering the settings menu.

TODO : In the future, maybe reload the settings in real time by having a file watcher. If the file is modified, then use an event to just reload the settings live.