Modified layout to move scrollbar inside of container
The previous layout was slightly wrong and caused some issues where the whole text was not visible when scrolling to the bottom.

The container box was a child of the scrollbar, rather than it being the other way around, which caused said issue, as well as cutting off rounded corners during scrolling and also rendering the scrollbar outside of the container. This issue has now been fixed by rendering the scrollbar and scrollable area inside of the container, as it should logically be done.