Buttons use bold text by default
By default, buttons in flutter use bold text. The recent isBold param that was added to the mpup text has made it so that there is a default value, which means that we now have to manually specify that the fonts are bold, otherwise they will be displayed normally again.