Modified buttons to have a default font size of 20
TODO : Modify the mpup buttons so that we can have any type of child widget within it, just as regular buttons. Either that or make an image button and a text button objects that return a baseline button with a pre-determined child object type.

That way, users will be capable of customizing the properties of their buttons, altough the differences between the default EvelevatedButton and the mpup button would start to disappear...