Added quick fix to allow the program to display help message for empty args
When no input path is given to the file processors (compiler or decompiler), the cmd help message for missing arguments is displayed.

For the GUI program, it would be more useful to see something like "you must input a file", or maybe even just disabling the button until a valid path is given, but for now, this will suffice.