Added file_selector plugin package to project
The file selector package offers a far better and more standard / system standard behaviour compliant interface for directory picking. This makes it better than the file_picker's directory picker.

Under the hood, on Windows and some other desktop platforms, the call to get a direcory picker actually freezes the main thread of the program. In the case of file picker, the user could still click on the frozen window and resize it, drag it around, etc... which lead to a terrible experience where it seemed like the app had crashed or something wrong was happening. The file_selector package on the other hand actually does properly handle this situation by doing what most windows applications do when directory picker dialogues appear, which is preventing users from clicking on the app window by throwing a blink animation and the error sound, preventing the freezing issue from being noticed.