Added aux property value loading function for permissive settings loading
With this new system, settings loading is now permissive. This means that if in a future update a new property is added, the settings json file does not fail to load, it simply loads the values that it can and ignores any that it could not find. The new fields will be generated properly the next time settings are modified by the program.

NOTE : For further support, a save format versioning scheme must be implemented, with different loading and saving functions.