Modified profiles menu to display internal profile name rather than directory name
The directory name is pretty much irrelevant and can be anything the user wants, as long as there are no collisions.

This is done like this to prevent issues with naming working different across different platforms and OSs, eg: on windows you can't name your profile anything that contains special symbols banned by the OS if we bind the profile name to the directory name. We also suffer the problem limiting the max length of the profile name depending on how long the prepended path is already.

In short, it just makes more sense for the actual name to be an internally stored property, and the profile directory name is just a subcategory made to further subdivide and organize the generated files and data.

As a note, I will admit that it does kind of make searching manually for an install more annoying, but we always have the button to open the selected isntall, so that is a non issue unless the user is not working with the mod manager.