Added try blocks to property reading and writing for profile data
Basic (and pretty shitty actually) error handling to prevent issues when loading and saving profile data.

The mechanism is extremely permissive, which means that we can technically attempt to load or save a profile and get a valid result which we could use despite the profile data not really existing on non-volatile storage.