Fixed oversight where the found entries was assigned to a temporary variable
The extra var keyword was creating a new "entries" variable that lived only within the scope of the if statement. This issue has now been resolved by properly assigning a list of found entries to the entries variable, allowing the rest of the program to make use of the found data.