Modified main to ensure existence of data directories
All of the app data is now going to be stored within a directory structure inside of the ./data directory, relative to the executable.

If the data dirs do not exist, they will be created automatically to ensure the app's correct functioning. The dir is required at least to store the settings file, which must always be located relative to the executable. All other paths are configurable.