Changed box display size
The larger box size is not a permanent fix, but it's ok for now.

Basically, this will allow a larger number of mods to be displayed on screen. Eventually tho, an issue will come up, which is the fact that a large enough number will not be capable of fitting on the fixed size container.

The solution is to use an intrinsic size container type, either that or scale the font size or container size according to the length of the number string itself.