Modified installs, mods and profiles menus to properly handle paths for directories even if they include dots in the dir name
The function pathName, due to historical reasons (and not wanting to rewrite half of the codebase's calls to pathName), has include extensions set to false by default.

This behaviour caused folders with names such as "Magicka 1.5.1.0" to get the "extension" .0 cropped out.

This is no longer the case.

NOTE : In the future, we may come across an issue in the profiles menu where profile dirs with an extension are not recognised properly. I have not added handling for that yet since dirs should be numbered automatically by mpupgui, but users can create their own folders and that could cause issues, but it is not critical as of now...